requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
diff-match-patch==20121119
Django==1.7.11
django-bootstrap3==6.2.1
django-celery==3.1.17
django-forms-bootstrap==3.0.1
django-guardian==1.4.4
django-import-export==0.4.5
django-jsonfield==0.9.19
django-simple-captcha==0.5.1
django-wkhtmltopdf==3.0.0
djangorestframework==3.3.3
easy-thumbnails==2.3
html2text==2016.4.2
jsonfield==0.9.23
kombu==3.0.35
lxml==3.6.0
Pillow==3.2.0
PyChart==1.39
pypinyin==0.12.1
pytz==2016.4
requests==2.10.0
six==1.10.0
tablib==0.11.2
xlrd==0.9.4
xlwt==1.0.0
